+++ src/queries/notifications/index.ts
@@ -4,13 +4,14 @@ import { useQuery } from '@tanstack/react-query'
 import { getNotification } from '@/services/notification'
 import type Notification from '@/types/notification'
 
-const useGetNotification = () => {
+const useGetNotification = ({ isLoggedIn }: { isLoggedIn: boolean }) => {
   return useQuery({
     queryKey: ['getNotification'],
     queryFn: async () => {
       const data: Notification[] = await getNotification()
       return data
     },
+    enabled: isLoggedIn,
   })
 }
